Apply review-round fixes: collapse cast helpers, rename folder, wire visible-linebreak toggle
- compileDOMRenderConfigOverrides: collapse merge3GetDOMSlot and
  ignoreNext3GetDOMSlot to type-cast aliases over merge3 / ignoreNext3
  per etrepum's suggestion, drop the variance comment.
- AutocompletePlugin folder renamed to AutocompleteExtension to match
  the extension conversion in 37100faa8.
- Add isVisibleLineBreak playground setting (mirrors isAutocomplete):
  toggle in Settings panel wires VisibleLineBreakExtension into the
  dynamic-config dependency list so etrepum's PR question about where
  to enable it has a concrete answer.
