+++ packages/lexical-playground/src/appSettings.ts
@@ -24,6 +24,7 @@ export const DEFAULT_SETTINGS = {
   isCollab: false,
   isMaxLength: false,
   isRichText: true,
+  isVisibleLineBreak: false,
   listStrictIndent: false,
   measureTypingPerf: false,
   selectionAlwaysOnDisplay: false,
